test(LoginForm): merge tests to one file
